Sprint 2 - Administrador - Visualizar um produto como é vendido na loja
Finalizado

+++ PharmaSW/src/main/java/br/com/pharmasw/api/modelo/Produto.java
@@ -23,7 +23,7 @@ public class Produto {
     private Double valor;
     private Double peso;
     private Integer quantidadeEstoque;
-    private Integer avaliacao;
+    private Double avaliacao;
 
     @Transient
     private String imagemPrincipal;
@@ -101,11 +101,11 @@ public class Produto {
         this.quantidadeEstoque = quantidadeEstoque;
     }
 
-    public Integer getAvaliacao() {
+    public Double getAvaliacao() {
         return avaliacao;
     }
 
-    public void setAvaliacao(Integer avaliacao) {
+    public void setAvaliacao(Double avaliacao) {
         this.avaliacao = avaliacao;
     }
